struct(T3): extract Game / GameAction / lifecycle into game.rs
