+++ src/game/mod.rs
@@ -1,9 +1,16 @@
+//! Rocker game simulation core.
+//!
+//! Exposes `Game` as the primary state machine, `GameAction` as the input
+//! command set, and submodules for simulation subsystems (band, player, world, etc.).
+
 mod actions;
 pub mod band;
 mod constants;
 mod economy;
 pub mod events;
 mod events_apply;
+#[allow(clippy::module_inception)]
+pub mod game;
 pub mod genre;
 pub mod music;
 pub mod player;
@@ -13,217 +20,30 @@ pub mod timeline;
 mod turn;
 pub mod world;
 
-use crate::data_loader::GameDataFiles;
-use crate::game::music::*; // For Song, Release, ReleaseType, MarketingCampaignType, ActiveMarketingCampaign
-use band::Band;
-pub use constants::{BREAK_WEEKS, PRESSING_TIERS};
-use events::EventManager;
-use player::Player;
-use rand::rngs::StdRng;
-use rand::{Rng, SeedableRng};
-use serde::{Deserialize, Serialize};
-use std::fs::File;
-use std::io::{Read, Write};
-use timeline::MusicTimeline;
-use world::{GameWorld, PotentialDealOffer};
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub enum GameAction {
-    LazeAround,
-    WriteSongs,
-    Practice,
-    RecordSingle { pressing: Option<usize> },
-    RecordAlbum { pressing: Option<usize> },
-    Gig(usize),
-    GoOnTour(usize),
-    TakeBreak,
-    VisitDoctor,
-    AcceptDeal(usize),
-    RejectDeal(usize),
-    AcceptSupportTour,
-    DeclineSupportTour,
-    StartMarketingCampaign(u32, MarketingCampaignType), // release_id, campaign_type
-    Quit,
-}
-
-/// An invitation from a bigger act to open on their tour.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SupportTourOffer {
-    pub host_band: String,
-    pub host_fame: u8,
-    pub weeks: u32,
-    pub pay: u32,
-    pub fame_gain: u8,
-    pub expires_week: u32,
-}
-
-/// The one deliberate use of ambient entropy: choosing a world seed when
-/// ROCKER_SEED doesn't dictate one. Every roll after this derives from it.
-fn default_seed() -> u64 {
-    rand::random::<u64>()
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct Game {
-    #[serde(default = "default_seed")]
-    pub world_seed: u64,
-    pub player: Player,
-    pub band: Band,
-    pub world: GameWorld,
-    pub events: EventManager,
-    pub timeline: MusicTimeline,
-    #[serde(skip)]
-    pub data_files: GameDataFiles,
-    pub pending_deal_offers: Vec<PotentialDealOffer>,
-    #[serde(default)]
-    pub pending_support_offer: Option<SupportTourOffer>,
-    #[serde(default)]
-    pub regional_fame: std::collections::HashMap<String, u8>,
-    /// Consecutive weeks with no public activity (no shows, nothing on sale).
-    #[serde(default)]
-    pub idle_streak: u32,
-    /// The last era-fit verdict the press reported on the band's genre
-    /// (-1 cold, 0 unremarkable, +1 hot) — the news speaks only on change.
-    #[serde(default)]
-    pub genre_trend_reported: i8,
-    pub week: u32,
-    pub game_over: bool,
-    pub next_song_id: u32,
-    pub next_release_id: u32,
-    pub just_released_music: Vec<Release>, // Stores releases for their initial sales window
-    /// Messages produced while processing the last turn, drained by the UI.
-    #[serde(skip)]
-    pub turn_log: Vec<String>,
-}
-
-impl Game {
-    pub fn new() -> Result<Self, Box<dyn std::error::Error>> {
-        GameDataFiles::validate_data_files()?;
-        let data_files = GameDataFiles::load()?;
-
-        let world_seed = std::env::var("ROCKER_SEED")
-            .ok()
-            .and_then(|s| s.parse::<u64>().ok())
-            .unwrap_or_else(default_seed);
-
-        let mut init_rng = StdRng::seed_from_u64(world_seed);
-        let world = GameWorld::new(&data_files, &mut init_rng);
-
-        let mut turn_log = Vec::new();
-        turn_log.push(format!("🌱 World Seed: {}", world_seed));
-
-        Ok(Self {
-            world_seed,
-            player: Player::default(),
-            band: Band::default(),
-            world,
-            events: EventManager::new(),
-            timeline: MusicTimeline::new(&data_files),
-            data_files,
-            pending_deal_offers: Vec::new(),
-            pending_support_offer: None,
-            regional_fame: std::collections::HashMap::new(),
-            idle_streak: 0,
-            genre_trend_reported: 0,
-            week: 1,
-            game_over: false,
-            next_song_id: 0,
-            next_release_id: 0,
-            just_released_music: Vec::new(),
-            turn_log,
-        })
-    }
-
-    pub(super) fn log(&mut self, message: impl Into<String>) {
-        self.turn_log.push(message.into());
-    }
-
-    pub fn take_turn_log(&mut self) -> Vec<String> {
-        std::mem::take(&mut self.turn_log)
-    }
-
-    /// The action-stream RNG for a given week: the same splitmix64 key
-    /// derivation the world stream uses in `advance_week_events`, applied to
-    /// the salted seed (see `ACTION_STREAM_SALT`). Derived on demand, never
-    /// stored — saves carry no RNG state, and a loaded game rolls exactly
-    /// what the unsaved one would have.
-    fn action_rng_for_week(&self, week: u64) -> StdRng {
-        let mut key = (self.world_seed ^ constants::ACTION_STREAM_SALT)
-            .wrapping_add(week)
-            .wrapping_mul(0x9E3779B97F4A7C15);
-        key = (key ^ (key >> 30)).wrapping_mul(0xBF58476D1CE4E5B8);
-        key = (key ^ (key >> 27)).wrapping_mul(0x94D049BB133111EB);
-        key ^= key >> 31;
-        StdRng::seed_from_u64(key)
-    }
-
-    /// Every roll made while resolving the current turn draws from this one
-    /// stream, in order: the action itself, then the week's random event,
-    /// then offer generation. Turn-consuming actions move the calendar, so
-    /// consecutive turns get fresh streams; the rare same-week paperwork
-    /// action (rejecting two deals in one sitting) rereads the week's stream,
-    /// which is deterministic and harmless.
-    pub(super) fn action_rng(&self) -> StdRng {
-        self.action_rng_for_week(self.week as u64)
-    }
-
-    pub fn initialize_player(
-        &mut self,
-        player_name: &str,
-        band_name: &str,
-        genre: genre::MusicGenre,
-    ) {
-        self.player.name = player_name.to_string();
-        self.band.name = band_name.to_string();
-        self.band.genre = genre;
-        self.player.money = 500; // Starting cash in 1970
-
-        // Bandmates are part of the seed's identity, like the scene itself.
-        let mut rng = self.action_rng_for_week(constants::SETUP_STREAM_WEEK);
-        self.band.members = vec![
-            band::BandMember {
-                name: self.data_files.random_band_member_name(&mut rng),
-                instrument: band::Instrument::Guitar,
-                skill: 25,
-                loyalty: 75,
-                drug_problem: false,
-            },
-            band::BandMember {
-                name: self.data_files.random_band_member_name(&mut rng),
-                instrument: band::Instrument::Bass,
-                skill: 20,
-                loyalty: 80,
-                drug_problem: false,
-            },
-            band::BandMember {
-                name: self.data_files.random_band_member_name(&mut rng),
-                instrument: band::Instrument::Drums,
-                skill: 30,
-                loyalty: 70,
-                drug_problem: false,
-            },
-        ];
-    }
-
-    pub fn save_game(&self, file_path: &str) -> Result<(), Box<dyn std::error::Error>> {
-        let json_string = serde_json::to_string_pretty(self)?;
-        let mut file = File::create(file_path)?;
-        file.write_all(json_string.as_bytes())?;
-        Ok(())
-    }
-
-    pub fn load_game(file_path: &str) -> Result<Self, Box<dyn std::error::Error>> {
-        let mut file = File::open(file_path)?;
-        let mut json_string = String::new();
-        file.read_to_string(&mut json_string)?;
-
-        let mut loaded_game: Game = serde_json::from_str(&json_string)?;
-
-        loaded_game.data_files = GameDataFiles::load()?;
-
-        Ok(loaded_game)
-    }
-}
-
 #[cfg(test)]
 mod tests;
+
+pub use constants::{BREAK_WEEKS, PRESSING_TIERS};
+pub use game::{Game, GameAction, SupportTourOffer};
+
+// Re-exports for submodules and external crates
+#[allow(unused_imports)]
+pub use crate::data_loader::GameDataFiles;
+#[allow(unused_imports)]
+pub use crate::game::music::{
+    ActiveMarketingCampaign, MarketingCampaignType, Release, ReleaseType,
+};
+#[allow(unused_imports)]
+pub use band::Band;
+#[allow(unused_imports)]
+pub use events::EventManager;
+#[allow(unused_imports)]
+pub use player::Player;
+#[allow(unused_imports)]
+pub use rand::rngs::StdRng;
+#[allow(unused_imports)]
+pub use rand::{Rng, SeedableRng};
+#[allow(unused_imports)]
+pub use timeline::MusicTimeline;
+#[allow(unused_imports)]
+pub use world::{GameWorld, PotentialDealOffer};
